Servicios de admin global y admin teatro terminados

+++ cine/src/main/java/co/edu/uniquindio/unicine/ServiciosImpl/AdminServicioImpl.java
@@ -3,7 +3,7 @@ package co.edu.uniquindio.unicine.ServiciosImpl;
 import co.edu.uniquindio.unicine.Entidades.*;
 import co.edu.uniquindio.unicine.Repo.*;
 import co.edu.uniquindio.unicine.Servicios.AdminServicio;
-import org.springframework.beans.factory.annotation.Autowired;
+import co.edu.uniquindio.unicine.Tipos.Tipo_Admin;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -23,13 +23,14 @@ public class AdminServicioImpl implements AdminServicio {
 
 
     public AdminServicioImpl(AdministrativoRepo administrativoRepo) {
+
         this.administrativoRepo = administrativoRepo;
     }
 
     @Override
     public Administrativo login(String cedula) throws Exception {
         Administrativo admin = administrativoRepo.obtenerPorCedula(cedula);
-        if(admin == null)
+        if(admin == null || !admin.getTipo().equals(Tipo_Admin.ADMINISTRADOR_GLOBAL))
         {
             throw new Exception("Los datos de autenticacion son incorrectos");
         }
@@ -37,10 +38,35 @@ public class AdminServicioImpl implements AdminServicio {
     }
 
     @Override
-    public Ciudad crearCiudad(Ciudad ciudad) {
+    public Ciudad crearCiudad(Ciudad ciudad)throws Exception {
+
         return ciudadRepo.save(ciudad);
     }
 
+    public void eliminarCiudad(Integer idCiudad)throws Exception{
+        Optional<Ciudad> guardado = ciudadRepo.findById(idCiudad);
+
+        if(guardado.isEmpty()){
+            throw new Exception("La ciudad que desea eliminar no existe");
+        }
+        ciudadRepo.delete(guardado.get());
+
+    }
+
+    public Ciudad actualizarCiudad(Ciudad ciudad) throws Exception{
+
+        Optional<Ciudad> guardado = ciudadRepo.findById(ciudad.getId());
+
+        if(guardado.isEmpty()) {
+            throw new Exception("La ciudad que desea actualizar no existe");
+        }
+        return ciudadRepo.save(ciudad);
+    }
+    public List<Ciudad> listarCiudad() {
+
+        return administrativoRepo.ListarCiudad();
+    }
+    //Gestionar teatro
     @Override
     public Teatro crearTeatro(Teatro teatro) throws Exception {
         return teatroRepo.save(teatro);
@@ -48,14 +74,25 @@ public class AdminServicioImpl implements AdminServicio {
 
     @Override
     public void eliminarTeatro(Integer idTeatro) throws Exception {
+        //buscamos por medio de la id y lo borramos
         Optional<Teatro> guardado = teatroRepo.findById(idTeatro);
+
+        if(guardado.isEmpty()){
+            throw new Exception("El teatro que desea eliminar no existe");
+        }
         teatroRepo.delete(guardado.get());
 
     }
 
     @Override
     public Teatro actualizarTeatro(Teatro teatro) throws Exception {
-        return null;
+        //buscamos en la lista de teatros por medio de la id
+        Optional<Teatro> guardado = teatroRepo.findById(teatro.getId());
+
+        if(guardado.isEmpty()) {
+            throw new Exception("El teatro que desea actualizar no existe");
+        }
+        return teatroRepo.save(teatro);
     }
 
     @Override
@@ -71,13 +108,22 @@ public class AdminServicioImpl implements AdminServicio {
     @Override
     public void eliminarCupon(Integer idCupon) throws Exception {
         Optional<Cupon> guardado = cuponRepo.findById(idCupon);
+
+        if(guardado.isEmpty()){
+            throw new Exception("El cupon que desea eliminar no existe");
+        }
         cuponRepo.delete(guardado.get());
 
     }
 
     @Override
     public Cupon actualizarCupon(Cupon cupon) throws Exception {
-        return null;
+        Optional<Cupon> guardado = cuponRepo.findById(cupon.getId());
+
+        if(guardado.isEmpty()) {
+            throw new Exception("El cupon que desea actualizar no existe");
+        }
+        return cuponRepo.save(cupon);
     }
 
     @Override
@@ -93,13 +139,22 @@ public class AdminServicioImpl implements AdminServicio {
     @Override
     public void eliminarConfiteria(Integer idConfiteria) throws Exception {
         Optional<Confiteria> guardado = confiteriaRepo.findById(idConfiteria);
+
+        if(guardado.isEmpty()){
+            throw new Exception("La confiteria que desea eliminar no existe");
+        }
         confiteriaRepo.delete(guardado.get());
 
     }
 
     @Override
     public Confiteria actualizarConfiteria(Confiteria confiteria) throws Exception {
-        return null;
+        Optional<Confiteria> guardado = confiteriaRepo.findById(confiteria.getId());
+
+        if(guardado.isEmpty()) {
+            throw new Exception("La confiteria que desea actualizar no existe");
+        }
+        return confiteriaRepo.save(confiteria);
     }
 
     @Override
@@ -115,13 +170,22 @@ public class AdminServicioImpl implements AdminServicio {
     @Override
     public void eliminarPelicula(Integer idPelicula) throws Exception {
         Optional<Pelicula> guardado = peliculaRepo.findById(idPelicula);
+
+        if(guardado.isEmpty()){
+            throw new Exception("La pelicula que desea eliminar no existe");
+        }
         peliculaRepo.delete(guardado.get());
 
     }
 
     @Override
     public Pelicula actualizarPelicula(Pelicula pelicula) throws Exception {
-        return null;
+        Optional<Pelicula> guardado = peliculaRepo.findById(pelicula.getId());
+
+        if(guardado.isEmpty()) {
+            throw new Exception("La pelicula que desea actualizar no existe");
+        }
+        return peliculaRepo.save(pelicula);
     }
 
     @Override
@@ -137,13 +201,22 @@ public class AdminServicioImpl implements AdminServicio {
     @Override
     public void eliminarAdminTeatro(String numCedula, Integer idAdmin) throws Exception {
         Optional<Administrativo> guardado = administrativoRepo.findById(idAdmin);
+
+        if(guardado.isEmpty()){
+            throw new Exception("El administrador que desea eliminar no existe");
+        }
         administrativoRepo.delete(guardado.get());
 
     }
 
     @Override
     public Administrativo actualizarAdminTeatro(Administrativo adminTeatro) throws Exception {
-        return null;
+        Optional<Administrativo> guardado = administrativoRepo.findById(adminTeatro.getId());
+
+        if(guardado.isEmpty()) {
+            throw new Exception("El Administrador de teatro que desea actualizar no existe");
+        }
+        return administrativoRepo.save(adminTeatro);
     }
 
     @Override
